requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
websockets==12.0
aiomysql==0.2.0
bcrypt==4.1.3
python-jose==3.3.0
redis==5.0.4
cryptography==42.0.5